requirements.txt
django==3.2.3
djangorestframework==3.12.4
django-filter==2.4.0
django-extensions==3.1.3
django-rest-auth==0.9.5
bcrypt==3.1.7
regex==2021.4.4
requests==2.23.0
validators==0.18.2
psycopg2-binary==2.9.1